Create Entities and relationships

+++ src/main/java/hcmute/puzzle/entities/Token.java
@@ -11,21 +11,21 @@ import java.util.Date;
 @AllArgsConstructor
 @Entity
 public class Token {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name = "id")
-    private int id;
+  @Id
+  @GeneratedValue(strategy = GenerationType.AUTO)
+  @Column(name = "id")
+  private long id;
 
-    @NonNull
-    @Column(name = "token")
-    private String token;
+  @NonNull
+  @Column(name = "token")
+  private String token;
 
-    @NonNull
-    @Column(name = "created_time")
-    private Date createDate;
+  @NonNull
+  @Column(name = "created_time")
+  private Date createDate;
 
-    @NonNull
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private UserEntity user;
+  @NonNull
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "user_id")
+  private UserEntity user;
 }
